refactor: seperate test & add dotenv
