+++ .env.example
@@ -1,0 +1,2 @@
+# for test
+NC_ACCESS_TOKEN=
