+++ package.json
@@ -35,7 +35,7 @@
   "scripts": {
     "build": "tsc",
     "dev": "nodemon",
-    "test": "tsx test.ts"
+    "test": "tsx test/index.ts"
   },
   "dependencies": {
     "events": "^3.3.0",
@@ -46,6 +46,7 @@
   "devDependencies": {
     "@types/node": "^22.5.1",
     "@types/ws": "^8.5.12",
+    "dotenv": "^16.4.5",
     "nodemon": "^3.1.4",
     "tsx": "^4.19.0",
     "typescript": "^5.5.4"
